% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDistT.R
\name{calcDistT}
\alias{calcDistT}
\title{Calculate total distance in given dataset}
\usage{
calcDistT(df)
}
\arguments{
\item{df}{A data frame with the first column indicating individual identity.}
}
\value{
Numeric. Total distance.
}
\description{
This function calculates centroid of all samples in a given dataset and sums
distances between the centroid and each sample. Euclidean distances are used.
}
\examples{
calcDistT(ANmodulation)

}
\seealso{
Other calcHM support function: \code{\link{calcDistW}},
  \code{\link{calcMeanVec}}
}
\concept{calcHM support function}
