% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDistW.R
\name{calcDistW}
\alias{calcDistW}
\title{Calculate average within individual distance}
\usage{
calcDistW(df)
}
\arguments{
\item{df}{A dataframe with the first column indicating individual identity.}
}
\value{
Numeric. Average within individual distance in dataset.
}
\description{
This function calculates centroid for each individual and sums distances of
samples from centroid for that particular individual. When the within
individual sum of distances is known for each individual, it calculates their
mean. Euclidean distances are used.
}
\examples{
calcDistW(ANmodulation)

}
