% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDS.R
\name{calcDS}
\alias{calcDS}
\title{Calculates discrimination score (DS)}
\usage{
calcDS(df)
}
\arguments{
\item{df}{A dataframe with the first column noting individual identity of the sample.}
}
\value{
Proportion of samples correctly classified by LDA in \code{df}.
}
\description{
This function will take the specified dataframe and will perform linear
discrimination analysis with leave-one-out crossvalidation and equal priors
for each individual (e.g., all priors will be set to 1/10 in case that
the dataset contains 10 individuals). Variables are not modified in any way;
scaling, centering, transformation of variables, or principal component
analysis, etc., if required, need to be done on dataset before calling this
function.
}
\examples{
calcDS(ANmodulation)

}
\seealso{
\code{\link{calcPIC}}, \code{\link{calcHS}}

Other individual identity metrics: \code{\link{calcF}},
  \code{\link{calcHM}}, \code{\link{calcHSngroups}},
  \code{\link{calcHSnpergroup}}, \code{\link{calcHSntot}},
  \code{\link{calcHSvarcomp}}, \code{\link{calcHS}},
  \code{\link{calcMI}}, \code{\link{calcPICbetweenmeans}},
  \code{\link{calcPICbetweentot}}, \code{\link{calcPIC}}
}
\concept{individual identity metrics}
