% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHM.R
\name{calcHM}
\alias{calcHM}
\title{Calculate information capacity (HM)}
\usage{
calcHM(df)
}
\arguments{
\item{df}{A dataframe with the first column indicating individual identity.}
}
\value{
Numeric value. Individual identity information capacity HM (in bits) in dataset.
}
\description{
This function calculates information capacity.
}
\examples{
calcHM(ANmodulation)

}
\seealso{
\code{\link{calcPIC}}, \code{\link{calcHS}}

Other individual identity metrics: \code{\link{calcDS}},
  \code{\link{calcF}}, \code{\link{calcHSngroups}},
  \code{\link{calcHSnpergroup}}, \code{\link{calcHSntot}},
  \code{\link{calcHSvarcomp}}, \code{\link{calcHS}},
  \code{\link{calcMI}}, \code{\link{calcPICbetweenmeans}},
  \code{\link{calcPICbetweentot}}, \code{\link{calcPIC}}
}
\concept{individual identity metrics}
