% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMI.R
\name{calcMI}
\alias{calcMI}
\title{Calculate Mutual information (MI)}
\usage{
calcMI(df)
}
\arguments{
\item{df}{A dataframe with the first column indicating individual identity.}
}
\value{
Numeric value of mutual information (in bits).
}
\description{
This function calculates Mutual information of actual and predicted
individual identities in the dataset. It uses Linear discrimination analysis
(MASS::lda) to predict individual identity. Settings for LDA are identical to
those used in \code{\link{calcDS}} function, i.e., LDA uses leave-one-out
cross-validation and priors are equal for each individual in dataset.
}
\examples{
calcMI(ANmodulation)

}
\seealso{
\code{\link{calcPIC}}, \code{\link{calcHS}}

Other individual identity metrics: \code{\link{calcDS}},
  \code{\link{calcF}}, \code{\link{calcHM}},
  \code{\link{calcHSngroups}},
  \code{\link{calcHSnpergroup}}, \code{\link{calcHSntot}},
  \code{\link{calcHSvarcomp}}, \code{\link{calcHS}},
  \code{\link{calcPICbetweenmeans}},
  \code{\link{calcPICbetweentot}}, \code{\link{calcPIC}}
}
\concept{individual identity metrics}
